requirements.txt
pynput.mouse
keyboard